feat: implement deferred token validation and server architecture refactoring (v1.5.0)
