--- CLAUDE.md
@@ -1,31 +1,0 @@
-# mcp-todoist Development Guidelines
-
-Auto-generated from all feature plans. Last updated: 2025-10-02
-
-## Active Technologies
-- (004-query-edit-completed)
-- TypeScript 5.x with Node.js 18+ + @modelcontextprotocol/sdk, axios (Todoist API client), zod (validation), existing TodoistApiService (005-bulk-action-on)
-- N/A (stateless operations, audit logging to console/file) (005-bulk-action-on)
-- TypeScript 5.x with Node.js 18+ + @modelcontextprotocol/sdk, existing TodoistApiService (axios-based) (006-more-mcp-compliance)
-- In-memory state (token validation cache, no persistent storage required) (006-more-mcp-compliance)
-
-## Project Structure
-```
-backend/
-frontend/
-tests/
-```
-
-## Commands
-# Add commands for 
-
-## Code Style
-: Follow standard conventions
-
-## Recent Changes
-- 006-more-mcp-compliance: Added TypeScript 5.x with Node.js 18+ + @modelcontextprotocol/sdk, existing TodoistApiService (axios-based)
-- 005-bulk-action-on: Added TypeScript 5.x with Node.js 18+ + @modelcontextprotocol/sdk, axios (Todoist API client), zod (validation), existing TodoistApiService
-- 004-query-edit-completed: Added
-
-<!-- MANUAL ADDITIONS START -->
-<!-- MANUAL ADDITIONS END -->
