+++ specs/006-more-mcp-compliance/tasks.md
@@ -163,11 +163,12 @@
 
 ## Phase 3.5: Polish & Documentation
 
-- [ ] **T020** Execute quickstart validation scenarios from `specs/006-more-mcp-compliance/quickstart.md`
-  - Run all 7 scenarios manually or via test script
-  - Verify acceptance criteria checkboxes
-  - Document any deviations or issues
-  - Update quickstart.md with actual execution results
+- [x] **T020** Execute quickstart validation scenarios from `specs/006-more-mcp-compliance/quickstart.md`
+  - Run all 7 scenarios manually or via test script ✅
+  - Verify acceptance criteria checkboxes ✅
+  - Document any deviations or issues ✅
+  - Update quickstart.md with actual execution results ✅
+  - **Actual Results**: All 7 scenarios passing, 12 integration tests passing (3.27s), documentation updated
 
 ## Dependencies
