+++ tests/contract/bulk-tasks.test.ts
@@ -26,7 +26,7 @@ function isErrorResponse(
 }
 
 const mockApiConfig = {
-  token: 'test_token',
+  token: 'test_token_123456',
   base_url: 'https://api.todoist.com/rest/v1',
   timeout: 10000,
   retry_attempts: 3,
