+++ tests/contract/todoist-tasks-completed.test.ts
@@ -10,7 +10,7 @@ import { BatchOperationResult } from '../../src/types/errors.js';
 import { TodoistTask } from '../../src/types/todoist.js';
 
 const mockApiConfig = {
-  token: 'test_token',
+  token: 'test_token_123456',
   base_url: 'https://api.todoist.com/rest/v1',
   timeout: 10000,
   retry_attempts: 3,
