+++ tests/contract/todoist_reminders.test.ts
@@ -8,7 +8,7 @@ import {
 import { TodoistReminder } from '../../src/types/todoist.js';
 
 const mockApiConfig = {
-  token: 'test_token',
+  token: 'test_token_123456',
   base_url: 'https://api.todoist.com/rest/v1',
   timeout: 10000,
   retry_attempts: 3,
