+++ tests/integration/batch-operations.test.ts
@@ -5,7 +5,7 @@ import { BatchOperationsService } from '../../src/services/batch.js';
 import { createInMemoryApiService } from '../helpers/inMemoryTodoistApiService.js';
 
 const mockApiConfig = {
-  token: 'test_token',
+  token: 'test_token_123456',
   base_url: 'https://api.todoist.com/rest/v1',
   timeout: 10000,
   retry_attempts: 3,
