+++ tests/integration/completed-tasks-workflow.test.ts
@@ -7,7 +7,7 @@ import { BatchOperationsService } from '../../src/services/batch.js';
 import { TodoistTask } from '../../src/types/todoist.js';
 
 const mockApiConfig = {
-  token: 'test_token',
+  token: 'test_token_123456',
   base_url: 'https://api.todoist.com/rest/v1',
   timeout: 10000,
   retry_attempts: 3,
