+++ tests/unit/natural-language-dates.test.ts
@@ -8,7 +8,7 @@ import { TodoistApiService } from '../../src/services/todoist-api.js';
 import { TodoistReminder } from '../../src/types/todoist.js';
 
 const mockApiConfig = {
-  token: 'test_token',
+  token: 'test_token_123456',
   base_url: 'https://api.todoist.com/rest/v1',
   timeout: 10000,
   retry_attempts: 3,
